% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_pair_align.R
\name{curve_pair_align}
\alias{curve_pair_align}
\title{Pairwise align two curves}
\usage{
curve_pair_align(beta1, beta2, mode = "O")
}
\arguments{
\item{beta1}{curve 1, provided as a matrix of dimensions \eqn{n \times T} for
\eqn{T}-dimensional curve evaluated on \eqn{n} sample points}

\item{beta2}{curve 2, provided as a matrix of dimensions \eqn{n \times T} for
\eqn{T}-dimensional curve evaluated on \eqn{n} sample points}

\item{mode}{Open (\code{"O"}) or Closed (\code{"C"}) curves}
}
\value{
a list containing \item{beta2n}{aligned curve 2 to 1}
\item{q2n}{aligned srvf 2 to 1}
\item{gam}{warping function}
\item{q1}{srvf of curve 1}
\item{beta1}{centered curve 1}
\item{beta2}{centered curve 2}
}
\description{
This function aligns to curves using Elastic Framework
}
\examples{
out <- curve_pair_align(beta[, , 1, 1], beta[, , 1, 5])
}
\references{
Srivastava, A., Klassen, E., Joshi, S., Jermyn, I., (2011). Shape analysis of elastic curves in euclidean spaces. Pattern Analysis and Machine Intelligence, IEEE Transactions on 33 (7), 1415-1428.
}
\keyword{alignment}
\keyword{srvf}
