% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_boxplot.R
\name{curvebox_data}
\alias{curvebox_data}
\title{Curve Boxplot Data}
\usage{
curvebox_data(align_median, alpha = 0.05, ka = 1)
}
\arguments{
\item{align_median}{object from \code{\link[=multivariate_karcher_mean]{multivariate_karcher_mean()}} of aligned curves using
the median}

\item{alpha}{quantile value (default=.05, i.e., 95\%)}

\item{ka}{scalar for outlier cutoff (default=1)}
}
\value{
Returns an \code{curvebox} object containing:
\itemize{
\item \code{median_y}: median curve
\item \code{Q1}: First quartile
\item \code{Q3}: Second quartile
\item \code{Q1a}: First quantile based on alpha
\item \code{Q3a}: Second quantile based on alpha
\item \code{minn}: minimum extreme curve
\item \code{maxx}: maximum extreme curve
\item \code{outlier_index}: indexes of outlier curves
\item \code{fmedian}: median curve
}
}
\description{
This function constructs the amplitude boxplot.
}
\references{
Xie, W., S. Kurtek, K. Bharath, and Y. Sun  (2016). "A geometric
approach to visualization of variability in functional data." Journal of
the American Statistical Association in press: 1-34.
}
\keyword{alignment}
\keyword{boxplot}
\keyword{internal}
\keyword{srvf}
