% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_prediction.R
\name{elastic.prediction}
\alias{elastic.prediction}
\title{Elastic Predicition from Regression Models}
\usage{
elastic.prediction(f, time, model, y = NULL, smooth_data = FALSE,
  sparam = 25)
}
\arguments{
\item{f}{matrix (\eqn{N} x \eqn{M}) of \eqn{M} functions with \eqn{N} samples}

\item{time}{vector of size \eqn{N} describing the sample points}

\item{model}{list describing model from elastic regression methods}

\item{y}{responses of test matrix f (default=NULL)}

\item{smooth_data}{smooth data using box filter (default = F)}

\item{sparam}{number of times to apply box filter (default = 25)}
}
\value{
Returns a list containing
\item{y_pred}{predicited values of f or probabilities depending on model}
\item{SSE}{sum of squared errors if linear}
\item{y_labels}{labels if logistic model}
\item{PC}{probability of classification if logistic}
}
\description{
This function performs prediction from an elastic regression model
 with phase-variablity
}
\keyword{(2014),}
\keyword{@references}
\keyword{A.,}
\keyword{Application}
\keyword{Classification,}
\keyword{D.,}
\keyword{Elastic}
\keyword{Electronic}
\keyword{Functional}
\keyword{J.}
\keyword{Journal}
\keyword{Logistic}
\keyword{Physiological}
\keyword{Regression}
\keyword{Signal}
\keyword{Srivastava,}
\keyword{Statistics}
\keyword{Tucker,}
\keyword{W.,}
\keyword{Wu,}
\keyword{alignment}
\keyword{of}
\keyword{regression}
\keyword{srvf}
\keyword{submitted.}
\keyword{to}
\keyword{with}

