% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_plot}
\alias{f_plot}
\title{Plot functional data}
\usage{
f_plot(t, f, ...)
}
\arguments{
\item{t}{A numeric vector of length \eqn{M} specifying the common grid on
which all curves \code{f} have been observed.}

\item{f}{A numeric matrix of shape \eqn{M \times N} specifying a sample of
\eqn{N} curves observed on a grid of size \eqn{M}.}
}
\description{
This function plots functional data on a time grid
}
