% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pns.R
\name{fastPNSs2e}
\alias{fastPNSs2e}
\title{Computes fast PNS sphere to euclidean}
\usage{
fastPNSs2e(spheredata, PNS)
}
\arguments{
\item{spheredata}{data on sphere}

\item{pns}{list of PNS structure}
}
\value{
EuclidData euclidean representation
}
\description{
Computes fast PNS sphere to euclidean
}
