\docType{package}
\name{fdasrvf}
\alias{fdasrvf}
\alias{fdasrvf-package}
\title{Elastic Functional Data Analysis}
\description{
  A library for functional data analysis using the square
  root velocity framework which performs pair-wise and
  group-wise alignment as well as modeling using functional
  component analysis
}
\references{
  Srivastava, A., Wu, W., Kurtek, S., Klassen, E., Marron,
  J. S., May 2011. Registration of functional data using
  fisher-rao metric, arXiv:1103.3817v2 [math.ST].
}

