% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_group_warp_bayes.R, R/function_mean_bayes.R
\name{function_mean_bayes}
\alias{function_mean_bayes}
\title{Bayesian Group Warping}
\usage{
function_mean_bayes(f, time, times = 5, group = 1:dim(f)[2],
  showplot = TRUE)

function_mean_bayes(f, time, times = 5, group = 1:dim(f)[2],
  showplot = TRUE)
}
\arguments{
\item{f}{matrix (\eqn{N} x \eqn{M}) of \eqn{M} functions with \eqn{N} samples}

\item{time}{sample points of functions}

\item{times}{number of subsample points to look at (default = 5)}

\item{group}{(default 1:dim(f)[2])}

\item{showplot}{shows plots of functions (default = T)}

\item{iter}{number of iterations (default = 150000)}

\item{powera}{(default 1)}

\item{tau}{(default ceil(times*.4))}

\item{gp}{(default seq(dim(input)[2]))}

\item{f}{matrix (\eqn{N} x \eqn{M}) of \eqn{M} functions with \eqn{N} samples}

\item{time}{sample points of functions}

\item{times}{number of subsample points to look at (default = 5)}

\item{showplot}{shows plots of functions (default = T)}
}
\value{
Returns a list containing \item{f0}{original functions}
\item{f_q}{f aligned using best}
\item{gam_q}{warping functions best}
\item{f_a}{f aligned using best mean}
\item{gam_a}{warping functions mean}
\item{qmn}{mean srsf}

Returns a list containing \item{distfamily}{dist matrix}
\item{match.matrix}{matches}
\item{position}{position}
\item{mu_5}{function mean}
\item{rtmatrix}{trmatrix}
\item{sumdist}{sumdist}
\item{qt.fitted}{aligned srsf functions}
\item{estimator}{estimator}
\item{estimator2}{estimator2}
\item{regfuncs}{registered functions}
}
\description{
This function aligns a set of functions using Bayesian SRSF framework

This function calculates karcher mean of functions using Bayesian SRSF framework
}
\examples{
\dontrun{
data("simu_data")
out = function_mean_bayes(simu_data$f, simu_data$time, iter=2)
}
\dontrun{
data("simu_data")
out = function_mean_bayes(simu_data$f, simu_data$time)
}
}
\references{
Cheng, W., Dryden, I. L., & Huang, X. (2016). Bayesian registration of functions and curves. Bayesian Analysis, 11(2), 447–475.

Cheng, W., Dryden, I. L., & Huang, X. (2016). Bayesian registration of functions and curves. Bayesian Analysis, 11(2), 447–475.
}
\keyword{alignment,}
\keyword{bayesian}
\keyword{srsf}

