% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactoring.R
\name{get_curve_centroid}
\alias{get_curve_centroid}
\title{Computes the centroid of a curve}
\usage{
get_curve_centroid(betafun)
}
\arguments{
\item{betafun}{A function that takes a numeric vector \eqn{s} of values in
\eqn{[0, 1]} as input and returns its values at \eqn{s}.}
}
\value{
A numeric vector of length \eqn{L} storing the centroid of the curve.
}
\description{
Computes the centroid of a curve
}
\examples{
betafun <- discrete2curve(beta[, , 1, 1])
get_curve_centroid(betafun)
}
