% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactoring.R
\name{get_identity_warping}
\alias{get_identity_warping}
\title{Computes the identity warping function}
\usage{
get_identity_warping()
}
\value{
A function that takes a numeric vector \eqn{s} of values in \eqn{[0,
  1]} as input and returns the values of the identity warping function at
\eqn{s}.
}
\description{
Computes the identity warping function
}
\examples{
get_identity_warping()
}
