% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactoring.R
\name{get_l2_hypersphere_distance}
\alias{get_l2_hypersphere_distance}
\title{Computes the geodesic distance between two SRVFs on the \eqn{L^2} hypersphere}
\usage{
get_l2_hypersphere_distance(q1fun, q2fun)
}
\arguments{
\item{q1fun}{A function that takes a numeric vector \eqn{s} of values in
\eqn{[0, 1]} as input and returns the values of the first SRVF at \eqn{s}.}

\item{q2fun}{A function that takes a numeric vector \eqn{s} of values in
\eqn{[0, 1]} as input and returns the values of the second SRVF at \eqn{s}.}
}
\value{
A numeric value storing the geodesic distance between the two SRVFs.
}
\description{
Computes the geodesic distance between two SRVFs on the \eqn{L^2} hypersphere
}
\examples{
q1 <- curve2srvf(beta[, , 1, 1])
q2 <- curve2srvf(beta[, , 1, 2])
q1p <- to_hypersphere(q1)
q2p <- to_hypersphere(q2)
get_l2_hypersphere_distance(q1p, q2p)
}
