% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_vel}
\alias{growth_vel}
\title{Berkeley Growth Velocity Dataset}
\format{
\subsection{\code{growth_vel}}{

A list with two components:
\itemize{
\item \code{f}: A numeric matrix of shape \eqn{69 \times 93} storing a sample of size
\eqn{N = 93} of curves evaluated on a grid of size \eqn{M = 69}.
\item \code{time}: A numeric vector of size \eqn{M = 69} storing the grid on which
the curves \code{f} have been evaluated.
}
}
}
\source{
???
}
\usage{
growth_vel
}
\description{
Combination of both boys and girls growth velocity from the Berkley dataset.
}
\keyword{datasets}
