% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{im}
\alias{im}
\title{Example Image Data set}
\format{
\subsection{\code{im}}{

A list with two components:
\itemize{
\item \code{I1}: A numeric matrix of shape \eqn{64 \times 64} storing the 1st image;
\item \code{I2}: A numeric matrix of shape \eqn{64 \times 64} storing the 2nd image.
}
}
}
\source{
???
}
\usage{
im
}
\description{
Contains two simulated images for registration.
}
\keyword{datasets}
