% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{inv_exp_map}
\alias{inv_exp_map}
\title{map square root of warping function to tangent space}
\usage{
inv_exp_map(Psi, psi)
}
\arguments{
\item{Psi}{vector describing psi function at center of tangent space}

\item{psi}{vector describing psi function to map to tangent space}
}
\value{
A numeric array of the same length as the input array \code{psi} storing the
shooting vector of \code{psi}
}
\description{
map square root of warping function to tangent space
}
\keyword{alignment}
\keyword{srvf}
