% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertGamma.R
\name{invertGamma}
\alias{invertGamma}
\title{Invert Warping Function}
\usage{
invertGamma(gam)
}
\arguments{
\item{gam}{vector of \eqn{N} samples}
}
\value{
Returns gamI inverted vector
}
\description{
This function calculates the inverse of gamma
}
\examples{
data("simu_warp")
out = invertGamma(simu_warp$gam[,1])
}
\references{
Srivastava, A., Wu, W., Kurtek, S., Klassen, E., Marron, J. S.,
 May 2011. Registration of functional data using fisher-rao metric,
 arXiv:1103.3817v2 [math.ST].

Tucker, J. D., Wu, W., Srivastava, A.,
 Generative Models for Function Data using Phase and Amplitude Separation,
 Computational Statistics and Data Analysis (2012), 10.1016/j.csda.2012.12.001.
}
\keyword{alignment}
\keyword{srvf}
