% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactoring.R
\name{matrix2curve}
\alias{matrix2curve}
\title{Converts a curve from matrix to functional data object}
\usage{
matrix2curve(beta)
}
\arguments{
\item{beta}{A numeric matrix of size \eqn{L \times M} specifying a curve on
an \eqn{L}-dimensional space observed on an evenly spaced grid of \eqn{[0,
 1]} of length \eqn{M}.}
}
\value{
A function that takes a numeric vector \eqn{s} of values in \eqn{[0,
  1]} as input and returns the values of the original curve at \eqn{s}.
}
\description{
Converts a curve from matrix to functional data object
}
\examples{
matrix2curve(beta[, , 1, 1])
}
