% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot Curve}
\usage{
plot_curve(beta, add = FALSE, ...)
}
\arguments{
\item{beta}{Array of sizes \eqn{n \times T} describing a
curve of dimension \eqn{n} evaluated on \eqn{T} points}

\item{add}{add to current plot (default = \code{TRUE})}

\item{...}{additional plotting parameters}
}
\value{
Return shape confidence intervals
}
\description{
This function plots open or closed curves
}
\keyword{bootstrap}
