% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.curve_pca.R
\name{predict.curve_pca}
\alias{predict.curve_pca}
\title{Elastic Prediction for curve PCA}
\usage{
\method{predict}{curve_pca}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class inheriting from "curve_pca"}

\item{newdata}{An optional matrix in which to look for functions with which to predict. If omitted, the original functions are used.}

\item{...}{additional arguments affecting the predictions produced}
}
\value{
Returns a matrix
\item{a}{principal coefficients}
}
\description{
This function performs projection of new curves on fPCA basis
}
\references{
Srivastava, A., Klassen, E., Joshi, S., Jermyn, I., (2011). Shape analysis of elastic curves in euclidean spaces. Pattern Analysis and Machine Intelligence, IEEE Transactions on 33 (7), 1415-1428.
}
\keyword{alignment}
\keyword{regression}
\keyword{srvf}
