% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.curve_pca.R, R/predict.hfpca.R
\name{predict.hfpca}
\alias{predict.hfpca}
\title{Elastic Prediction for curve PCA}
\usage{
\method{predict}{hfpca}(object, newdata = NULL, ...)

\method{predict}{hfpca}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class inheriting from "horizFPCA"}

\item{newdata}{An optional matrix in which to look for functions with which to predict. If omitted, the original functions are used.}

\item{...}{additional arguments affecting the predictions produced}
}
\value{
Returns a matrix
\item{a}{principal coefficients}

Returns a matrix
\item{a}{principal coefficients}
}
\description{
This function performs projection of new curves on fPCA basis

This function performs projection of new functions on fPCA basis
}
\references{
Srivastava, A., Klassen, E., Joshi, S., Jermyn, I., (2011). Shape analysis of elastic curves in euclidean spaces. Pattern Analysis and Machine Intelligence, IEEE Transactions on 33 (7), 1415-1428.

Tucker, J. D., Wu, W., Srivastava, A.,
Generative Models for Function Data using Phase and Amplitude Separation,
Computational Statistics and Data Analysis (2012), 10.1016/j.csda.2012.12.001.
}
\keyword{alignment}
\keyword{regression}
\keyword{srvf}
