% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.jfpca.R
\name{predict.jfpca}
\alias{predict.jfpca}
\title{Elastic Prediction for functional PCA}
\usage{
\method{predict}{jfpca}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class inheriting from "jointFPCA"}

\item{newdata}{An optional matrix in which to look for functions with which to predict. If omitted, the original functions are used.}

\item{...}{additional arguments affecting the predictions produced}
}
\value{
Returns a matrix
\item{a}{principle coefficients}
}
\description{
This function performs projection of new functions on fPCA basis
}
\references{
Tucker, J. D., Wu, W., Srivastava, A.,
Generative Models for Function Data using Phase and Amplitude Separation,
Computational Statistics and Data Analysis (2012), 10.1016/j.csda.2012.12.001.
}
\keyword{alignment}
\keyword{regression}
\keyword{srvf}
