% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{psi_to_gam}
\alias{psi_to_gam}
\title{map hilbert sphere to warping function}
\usage{
psi_to_gam(v)
}
\arguments{
\item{psi}{Either a numeric vector of a numeric matrix or a numeric array
specifying the shooting vectors}
}
\value{
A numeric array of the same shape as the input array \code{psi} storing the
warping functions \code{psi}.
}
\description{
map hilbert sphere to warping function
}
\keyword{alignment}
\keyword{srvf}
