% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplecurve.R
\name{resamplecurve}
\alias{resamplecurve}
\title{Resample Curve}
\usage{
resamplecurve(x, N = 100)
}
\arguments{
\item{x}{matrix defining curve (n,T)}

\item{N}{Number of samples to re-sample curve, N usually is > T}
}
\value{
xn matrix defining resampled curve
}
\description{
This function resamples a curve to a number of points
}
\examples{
data("curve_data")
xn = resamplecurve(curve_data$x,100)
}
\references{
Srivastava, A., Klassen, E., Joshi, S., Jermyn, I., (2011). Shape analysis of elastic curves in euclidean spaces. Pattern Analysis and Machine Intelligence, IEEE Transactions on 33 (7), 1415-1428.
}
\keyword{alignment}
\keyword{srvf}

