% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_pca}
\alias{rotation_pca}
\title{Rotation Principal Component Analysis}
\usage{
rotation_pca(align_data, no = 3, var_exp = NULL)
}
\arguments{
\item{align_data}{fdacurve object from \link{multivariate_karcher_mean} of aligned data}

\item{no}{number of principal components to extract}

\item{var_exp}{compute no based on value percent variance explained (example: 0.95)
will override \code{no}}
}
\value{
Returns a rotpca object containing
\item{latent}{latent values}
\item{coef}{coefficients}
\item{U}{eigenvectors}
}
\description{
This function calculates functional principal component analysis
on rotation data from
}
