% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Distributed Gaussian Peak Dataset}
\format{
\subsection{\code{toy_data}}{

A list with two components:
\itemize{
\item \code{f}: A numeric matrix of shape \eqn{101 \times 29} storing a sample of size
\eqn{N = 29} of curves evaluated on a grid of size \eqn{M = 101}.
\item \code{time}: A numeric vector of size \eqn{M = 101} storing the grid on which
the curves \code{f} have been evaluated.
}
}
}
\usage{
toy_data
}
\description{
A functional dataset where the individual functions are given by a Gaussian
peak with locations along the \eqn{x}-axis. The variables are as follows: f
containing the 29 functions of 101 samples and time which describes the
sampling.
}
\keyword{datasets}
