\docType{data}
\name{toy_warp}
\alias{toy_warp}
\title{Aligned Distrubted Gaussian Peak Dataset}
\format{A list which contains the outputs of the time_warping function}
\description{
  A functional dataset where the individual functions are
  given by a gaussian peak with locations along the
  \eqn{x}-axis. The variables are as follows: f containing
  the 29 functions of 101 samples and time which describes
  the sampling which as been aligned
}
\keyword{datasets}

