% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{v_to_curve}
\alias{v_to_curve}
\title{map shooting vector to curve at mean}
\usage{
v_to_curve(v, mu, mode = "O", scale = 1)
}
\arguments{
\item{v}{Either a numeric vector of a numeric matrix or a numeric array
specifying the shooting vectors}

\item{mu}{vector describing the mean}

\item{mode}{A character string specifying whether the input curves should be
considered open (\code{mode == "O"}) or closed (\code{mode == "C"}). Defaults to
\code{"O"}.}

\item{scale}{original scale of curve}
}
\value{
A numeric array of the same shape as the input array \code{v} storing the
curves of \code{v}.
}
\description{
map shooting vector to curve at mean
}
\keyword{alignment}
\keyword{srvf}
