% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{v_to_gam}
\alias{v_to_gam}
\title{map shooting vector to warping function at identity}
\usage{
v_to_gam(v)
}
\arguments{
\item{v}{Either a numeric vector of a numeric matrix or a numeric array
specifying the shooting vectors}
}
\value{
A numeric array of the same shape as the input array \code{v} storing the
shooting vectors of \code{v} obtained via finite differences.
}
\description{
map shooting vector to warping function at identity
}
\keyword{alignment}
\keyword{srvf}
