% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactoring.R
\name{vector2warping}
\alias{vector2warping}
\title{Converts a warping function from vector to functional data object}
\usage{
vector2warping(gam)
}
\arguments{
\item{gam}{A numeric vector of length \eqn{M} specifying a warping function
on an evenly spaced grid of \eqn{[0, 1]} of length \eqn{M}.}
}
\value{
A function that takes a numeric vector \eqn{s} of values in \eqn{[0,
  1]} as input and returns the values of the original warping function at
\eqn{s}.
}
\description{
Converts a warping function from vector to functional data object
}
\examples{
vector2warping(toy_warp$gam[, 1])
}
