% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdasrvf-package.r
\docType{data}
\name{beta}
\alias{beta}
\title{MPEG7 Curve Dataset}
\format{an array of shape (2,100,65,20)}
\usage{
data("mpeg7")
}
\description{
Contains the MPEG7 curve data set which is 20 curves in 65 classes. The array
is structured with dimension (2,100,65,20)
}
\keyword{datasets}
