% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_to_q.R
\name{curve_to_q}
\alias{curve_to_q}
\title{Convert to SRVF space}
\usage{
curve_to_q(beta)
}
\arguments{
\item{beta}{array describing curve (n,T)}
}
\value{
q array describing srvf
}
\description{
This function converts curves to SRVF
}
\examples{
data("mpeg7")
q = curve_to_q(beta[,,1,1])
}
\references{
Srivastava, A., Klassen, E., Joshi, S., Jermyn, I., (2011). Shape analysis of elastic curves in euclidean spaces. Pattern Analysis and Machine Intelligence, IEEE Transactions on 33 (7), 1415-1428.
}
\keyword{alignment}
\keyword{srvf}
