% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic.depth.R
\name{elastic.depth}
\alias{elastic.depth}
\title{Calculates elastic depth}
\usage{
elastic.depth(f, time, lambda = 0, parallel = FALSE)
}
\arguments{
\item{f}{matrix of N function of M time points (MxN)}

\item{time}{sample points of functions}

\item{lambda}{controls amount of warping (default = 0)}

\item{parallel}{run computation in parallel (default = T)}
}
\value{
Returns a list containing \item{amp}{amplitude depth}
\item{phase}{phase depth}
}
\description{
This functions calculates the elastic depth between set of functions
}
\examples{
data("simu_data")
depths <- elastic.depth(simu_data$f[,1:4],simu_data$time)
}
\references{
T. Harris, J. D. Tucker, B. Li, and L. Shand, "Elastic depths for detecting shape anomalies in functional data," Technometrics, 10.1080/00401706.2020.1811156, 2020.
}
\concept{srvf alignment}
\keyword{depth}
