% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_srvf.R
\name{f_to_srvf}
\alias{f_to_srvf}
\title{Convert to SRSF}
\usage{
f_to_srvf(f, time)
}
\arguments{
\item{f}{matrix of functions}

\item{time}{time}
}
\value{
q matrix of SRSFs
}
\description{
This function converts functions to srsf
}
\examples{
data("simu_data")
q <- f_to_srvf(simu_data$f,simu_data$time)
}
\references{
Srivastava, A., Wu, W., Kurtek, S., Klassen, E., Marron, J. S.,
 May 2011. Registration of functional data using fisher-rao metric,
 arXiv:1103.3817v2 [math.ST].

Tucker, J. D., Wu, W., Srivastava, A.,
 Generative Models for Function Data using Phase and Amplitude Separation,
 Computational Statistics and Data Analysis (2012), 10.1016/j.csda.2012.12.001.
}
\keyword{alignment}
\keyword{srsf}
