% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Gradient using finite differences}
\usage{
gradient(f, binsize)
}
\arguments{
\item{f}{vector with \eqn{N} samples}

\item{binsize}{scalar of time samples}
}
\value{
g vector with \eqn{N} samples which is the gradient of f
}
\description{
This function takes the gradient of f using finite differences
}
\examples{
data("simu_data")
out = gradient(simu_data$f[,1],mean(diff(simu_data$time)))
}
\keyword{alignment}
\keyword{srvf}
