% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdasrvf-package.r
\docType{data}
\name{growth_vel}
\alias{growth_vel}
\title{Berkley Growth Velocity Dataset}
\format{A list which contains f and time}
\usage{
data("growth_vel")
}
\description{
Combination of both boys and girls growth velocity from the Berkley Dataset
}
\keyword{datasets}
