% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdasrvf-package.r
\docType{data}
\name{im}
\alias{im}
\title{Example Image Data set}
\format{
a list containing two images of dimension  (64,64)
}
\usage{
data("image")
}
\description{
Contains two simulated images for registration
}
\keyword{datasets}
