% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{psi_to_gam}
\alias{psi_to_gam}
\title{map Hilbert sphere to warping function}
\usage{
psi_to_gam(psi)
}
\arguments{
\item{psi}{Either a numeric vector of a numeric matrix or a numeric array
specifying the points on the Hilbert Sphere}
}
\value{
A numeric array of the same shape as the input array \code{psi} storing the
warping function obtained via finite integration
}
\description{
map Hilbert sphere to warping function
}
\keyword{alignment}
\keyword{srvf}
