% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_warping.R
\name{time_warping}
\alias{time_warping}
\title{Alignment of univariate functional data}
\usage{
time_warping(
  f,
  time,
  lambda = 0,
  penalty_method = c("roughness", "geodesic", "norm"),
  centroid_type = c("mean", "median"),
  center_warpings = TRUE,
  show_plot = TRUE,
  smooth_data = FALSE,
  sparam = 25L,
  parallel = FALSE,
  optim_method = c("DP", "DP2", "RBFGS"),
  max_iter = 20L
)
}
\arguments{
\item{f}{A numeric matrix of shape \eqn{M \times N} specifying a sample of
\eqn{N} curves observed on a grid of size \eqn{M}.}

\item{time}{A numeric vector of length \eqn{M} specifying the common grid on
which all curves \code{f} have been observed.}

\item{lambda}{A numeric value specifying the elasticity. Defaults to \code{0.0}.}

\item{penalty_method}{A string specifying the penalty term used in the
formulation of the cost function to minimize for alignment. Choices are
\code{"roughness"} which uses the norm of the second derivative, \code{"geodesic"}
which uses the geodesic distance to the identity and \code{"norm"} which uses
the Euclidean distance to the identity. Defaults to \code{"roughness"}.}

\item{centroid_type}{A string specifying the type of centroid to align to.
Choices are \code{"mean"} or \code{"median"}. Defaults to \code{"mean"}.}

\item{center_warpings}{A boolean specifying whether to center the estimated
warping functions. Defaults to \code{TRUE}.}

\item{show_plot}{A boolean specifying whether to display plots of original
and wrapped functions. Defaults to \code{TRUE}.}

\item{smooth_data}{A boolean specifying whether to smooth curves using a box
filter. Defaults to \code{FALSE}.}

\item{sparam}{An integer value specifying the number of times to apply the
box filter. Defaults to \code{25L}. This is used only when \code{smooth_data = TRUE}.}

\item{parallel}{A boolean specifying whether to run calculations in parallel.
Defaults to \code{FALSE}.}

\item{optim_method}{A string specifying the algorithm used for optimization.
Choices are \code{"DP"}, \code{"DP2"} and \code{"RBFGS"}. Defaults to \code{"DP"}.}

\item{max_iter}{An integer value specifying the maximum number of iterations.
Defaults to \code{20L}.}
}
\value{
An object of class \code{fdawarp} which is a list with the following
components:
\itemize{
\item \code{f0}: original functions;
\item \code{fn}: aligned functions - matrix (\eqn{N} x \eqn{M}) of \eqn{M} functions
with \eqn{N} samples;
\item \code{qn}: aligned SRSFs - similar structure to fn;
\item \code{q0}: original SRSF - similar structure to fn;
\item \code{fmean}: function mean or median - vector of length \eqn{N};
\item \code{mqn}: SRSF mean or median - vector of length \eqn{N};
\item \code{gam}: warping functions - similar structure to fn;
\item \code{orig.var}: Original Variance of Functions;
\item \code{amp.var}: Amplitude Variance;
\item \code{phase.var}: Phase Variance;
\item \code{qun}: Cost function value.
}
}
\description{
This function aligns a collection of \eqn{1}-dimensional curves that are
observed on the same grid.
}
\examples{
\dontrun{
  out <- time_warping(simu_data$f, simu_data$time)
}
}
\references{
Srivastava, A., Wu, W., Kurtek, S., Klassen, E., Marron, J. S.,
May 2011. Registration of functional data using Fisher-Rao metric,
arXiv:1103.3817v2.

Tucker, J. D., Wu, W., Srivastava, A., Generative models for
functional data using phase and amplitude Separation, Computational
Statistics and Data Analysis (2012), 10.1016/j.csda.2012.12.001.
}
\keyword{alignment}
\keyword{srsf}
