% Generated by roxygen2 (4.0.1): do not edit by hand
\name{time_warping}
\alias{time_warping}
\title{Group-wise function alignment}
\usage{
time_warping(f, time, lambda = 0, method = "mean", showplot = TRUE,
  smooth_data = FALSE, sparam = 25, parallel = FALSE, cores = 2)
}
\arguments{
\item{f}{matrix (\eqn{N} x \eqn{M}) of \eqn{M} functions with \eqn{N} samples}

\item{time}{vector of size \eqn{N} describing the sample points}

\item{lambda}{controls the elasticity (default = 0)}

\item{method}{warp and caluclate to Karcher Mean or Median (options = "mean"
or "median", default = "mean")}

\item{showplot}{shows plots of functions (default = T)}

\item{smooth_data}{smooth data using box filter (default = F)}

\item{sparam}{number of times to apply box filter (default = 25)}

\item{parallel}{enable parallel mode using \code{\link{foreach}} and
\code{doParallel} pacakge}

\item{cores}{set number of cores to use with \code{doParallel} (default = 2)}
}
\value{
Returns a list containing \item{f0}{original functions}
\item{fn}{aligned functions - matrix (\eqn{N} x \eqn{M}) of \eqn{M} functions with \eqn{N} samples}
\item{qn}{aligned srvfs - similar structure to fn}
\item{q0}{original srvf - similar structure to fn}
\item{fmean}{function mean or median - vector of length \eqn{N}}
\item{mqn}{srvf mean or median - vector of length \eqn{N}}
\item{gam}{warping functions - similar structure to fn}
\item{orig.var}{Original Variance of Functions}
\item{amp.var}{Amplitude Variance}
\item{phase.var}{Phase Variance}
\item{qun}{Cost Function Value}
}
\description{
This function aligns a collection of functions using the elastic square-root
velocity (srvf) framework.
}
\examples{
data("simu_data")
out = time_warping(simu_data$f,simu_data$time)
}
\references{
Srivastava, A., Wu, W., Kurtek, S., Klassen, E., Marron, J. S.,
 May 2011. Registration of functional data using fisher-rao metric,
 arXiv:1103.3817v2 [math.ST].

Tucker, J. D., Wu, W., Srivastava, A.,
 Generative Models for Function Data using Phase and Amplitude Separation,
 Computational Statistics and Data Analysis (2012), 10.1016/j.csda.2012.12.001.
}
\keyword{alignment}
\keyword{srvf}

