% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactoring.R
\name{to_hypersphere}
\alias{to_hypersphere}
\title{Projects an SRVF onto the \eqn{L^2} hypersphere}
\usage{
to_hypersphere(qfun)
}
\arguments{
\item{qfun}{A function that takes a numeric vector \eqn{s} of values in
\eqn{[0, 1]} as input and returns the values of the SRVF at \eqn{s}.}
}
\value{
A function that takes a numeric vector \eqn{s} of values in \eqn{[0,
  1]} as input and returns the values of the SRVF projected onto the
\eqn{L^2} hypersphere.
}
\description{
Projects an SRVF onto the \eqn{L^2} hypersphere
}
\examples{
q <- curve2srvf(beta[, , 1, 1])
to_hypersphere(q)
}
