% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdasrvf-package.r
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Distributed Gaussian Peak Dataset}
\format{A list which contains f and time}
\usage{
data("toy_data")
}
\description{
A functional dataset where the individual functions are given by a Gaussian
peak with locations along the \eqn{x}-axis. The variables are as follows:
f containing the 29 functions of 101 samples and time which describes the
sampling
}
\keyword{datasets}
