\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Distrubted Gaussian Peak Dataset}
\format{A list which contains f and time}
\description{
  A functional dataset where the individual functions are
  given by a gaussian peak with locations along the
  \eqn{x}-axis. The variables are as follows: f containing
  the 29 functions of 101 samples and time which describes
  the sampling
}
\keyword{datasets}

