% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_warp}
\alias{toy_warp}
\title{Aligned Distributed Gaussian Peak Dataset}
\format{
\subsection{\code{toy_warp}}{

A list which contains the output of the \code{\link[=time_warping]{time_warping()}} function applied on
the data set \code{toy_data}.
}
}
\source{
???
}
\usage{
toy_warp
}
\description{
A functional dataset where the individual functions are given by a Gaussian
peak with locations along the \eqn{x}-axis. The variables are as follows: f
containing the 29 functions of 101 samples and time which describes the
sampling which as been aligned.
}
\keyword{datasets}
